Refactoring
Extracting Button, layout, Header, Types and InputField to separated files
