+++ client/src/components/EventForm/EventFormStyle.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+const EventFormStyles = styled.form`
+  min-width: 350px;
+  border: 1px solid #ccc;
+  padding: 30px;
+  display: flex;
+  flex-direction: column;
+`;
+
+export default EventFormStyles;
