+++ client/src/components/Layout/Container.tsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+
+const Container = styled.div`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+`;
+
+export default Container;
