+++ client/src/components/common/Header.tsx
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+const Header = styled.h1`
+  font-family: "Permanent Marker", cursive;
+  text-align: center;
+`;
+export default Header;
