+++ client/src/components/common/InputField.tsx
@@ -1,0 +1,9 @@
+import styled from "styled-components";
+import TextField from "@mui/material/TextField";
+
+const InputField = styled(TextField)`
+  margin: 1rem 0;
+  font-family: "Permanent Marker", cursive;
+`;
+
+export default InputField;
