+++ client/src/types/EventFormTypes.ts
@@ -1,0 +1,13 @@
+export interface EventFormBody {
+  firstName: string;
+  lastName: string;
+  email: string;
+  eventDate: string;
+}
+
+export interface ValidationErrors {
+  firstName?: string;
+  lastName?: string;
+  email?: string;
+  eventDate?: string;
+}
